gemini-streamlit-cloudrun/requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-logging
google-cloud-bigquery
db_dtypes
PyPDF2
backoff
tqdm